move new cache implementation to allensdk/api/cloud_cache/
