+++ allensdk/api/cloud_cache/cloud_cache.py
@@ -7,11 +7,11 @@ import boto3
 from botocore import UNSIGNED
 from botocore.client import Config
 from allensdk.internal.core.lims_utilities import safe_system_path
-from allensdk.brain_observatory.visual_behavior_cache.manifest import Manifest
-from allensdk.brain_observatory.visual_behavior_cache.file_attributes import CacheFileAttributes  # noqa: E501
-from allensdk.brain_observatory.visual_behavior_cache.utils import file_hash_from_path  # noqa: E501
-from allensdk.brain_observatory.visual_behavior_cache.utils import bucket_name_from_uri  # noqa: E501
-from allensdk.brain_observatory.visual_behavior_cache.utils import relative_path_from_uri  # noqa: E501
+from allensdk.api.cloud_cache.manifest import Manifest
+from allensdk.api.cloud_cache.file_attributes import CacheFileAttributes  # noqa: E501
+from allensdk.api.cloud_cache.utils import file_hash_from_path  # noqa: E501
+from allensdk.api.cloud_cache.utils import bucket_name_from_uri  # noqa: E501
+from allensdk.api.cloud_cache.utils import relative_path_from_uri  # noqa: E501
 
 
 class CloudCache(object):
