+++ allensdk/api/cloud_cache/file_attributes.py
